fix gitignore

+++ include/util/Point.h
@@ -2,6 +2,6 @@
 #include <cstdint>
 
 struct Point {
-    int32_t x {};
-    int32_t y {};
+    float x {};
+    float y {};
 };
